requirements.txt
anndata==0.7.8
coverage==6.1.2
genetools==0.6.0
h5py==3.6.0
igraph==0.9.9
ipykernel==5.3.4
joblib==1.1.0
jupytext==1.11.5
louvain==0.7.1
matplotlib==3.3.2
numpy==1.19.1
pandas==1.2.5
pre-commit==2.15.0
pytest==6.2.2
pytest-cov==3.0.0
python-slugify==5.0.2
scanpy==1.8.2
scikit-learn==1.0
scipy==1.6.2
seaborn==0.11.2
summarynb==0.1.4
